+++ src/test/java/de/alpharogroup/comparators/LocaleComparatorTest.java
@@ -37,11 +37,11 @@ import org.testng.annotations.Test;
 public class LocaleComparatorTest
 {
 
-	/** For use of the expected result. */
-	boolean expected;
-
 	/** For use of the result of the comparator. */
 	int actual;
+
+	/** For use of the expected result. */
+	boolean expected;
 
 	/**
 	 * Test method for {@link LocaleComparator#compare(Locale, Locale)}