+++ src/test/java/de/alpharogroup/comparators/StringLengthComparatorTest.java
@@ -39,14 +39,14 @@ import org.testng.annotations.Test;
 public class StringLengthComparatorTest
 {
 
-	/** For use of the expected result. */
-	boolean expected;
-
 	/** For use of the result of the comparator. */
 	int actual;
 
 	/** The comparator. */
 	Comparator<String> comparator;
+
+	/** For use of the expected result. */
+	boolean expected;
 
 	/**
 	 * Test method for {@link StringLengthComparator#compare(String, String)}